# docs/swagger.yaml
definitions:
...
  api.UserSexnovelHistroyAddRequest:
    properties:
      sexnovel_id:
        type: integer
    type: object
  api.UserSexnovelStarAddRequest:
    properties:
      sexnovel_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /api//sexnovel/chapte/list:
    get:
      parameters:
      - in: query
        name: id
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: pid
        type: integer
      - in: query
        name: title
        type: string
      responses: {}
      summary: 小说章节列表
      tags:
      - 小说
...
  /api/sexnovel/content/info:
    get:
      consumes:
      - application/json
      description: 返回小说内容数据
      parameters:
      - description: 小说内容的章节id
        in: query
        name: id
        required: true
        type: string
      responses: {}
      summary: 小说内容
      tags:
      - 小说
  /api/sexnovel/info:
    get:
      consumes:
      - application/json
      description: 返回小说详情数据
      parameters:
      - description: 小说详情的id
        in: query
        name: id
        required: true
        type: string
      responses: {}
      summary: 小说详情
      tags:
      - 小说
...
  /api/sexnovel/list:
    get:
      parameters:
      - in: query
        name: appId
        type: integer
      - in: query
        name: id
        type: integer
      - in: query
        name: labelsId
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: title
        type: string
      - in: query
        name: top
        type: integer
      - in: query
        name: typeid
        type: integer
      responses: {}
      summary: 小说列表
      tags:
      - 小说
...
  /api/user/sexnovel/history/add:
    put:
      consumes:
      - application/json
      description: 创建小说观看记录，单用户与单小说唯一。
      parameters:
      - description: 参数列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UserSexnovelHistroyAddRequest'
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 创建小说观看记录
      tags:
      - 小说
  /api/user/sexnovel/history/delete:
    delete:
      consumes:
      - application/json
      description: 删除小说观看记录 id 为观看记录id:'1,2,3' id=0 则为 删除所有
      parameters:
      - description: 删除观看历史的id
        in: query
        name: id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 删除小说观看记录
      tags:
      - 小说
  /api/user/sexnovel/history/list:
    get:
      consumes:
      - application/json
      description: 创建小说观看记录，单用户与单小说唯一。
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取小说观看记录
      tags:
      - 小说
  /api/user/sexnovel/star/add:
    put:
      consumes:
      - application/json
      description: 创建点赞收藏记录
      parameters:
      - description: 参数列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UserSexnovelStarAddRequest'
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 创建点赞收藏记录
      tags:
      - 小说
  /api/user/sexnovel/star/delete:
    delete:
      consumes:
      - application/json
      description: 创建小说点赞收藏记录，单用户与单小说唯一。
      parameters:
      - description: 删除点赞收藏的 sexnovel id
        in: query
        name: sexnovel_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 删除点赞收藏列表
      tags:
      - 小说
  /api/user/sexnovel/star/isLike:
    get:
      description: 返回isLike
      parameters:
      - description: 小说的id
        in: query
        name: sexnovel_id
        required: true
        type: string
      responses: {}
      summary: 是否点赞收藏
      tags:
      - 小说
  /api/user/sexnovel/star/list:
    get:
      consumes:
      - application/json
      description: 获取某个用户的点赞收藏记录
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: type
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取点赞收藏记录
      tags:
      - 小说
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"